fix load_optional test with Module::Runtime
It doesn't seem like the existing test could have ever worked.

use_package_optimistically returns the module name passed to it if
either the module is loaded, or if it is not found. It throws an error
if the module was found but had a compile error. It will never return a
false value. It can't be used on its own to test if a module was loaded.

Update the test to check the module's %INC entry after trying to load it
with use_package_optimistically. This seems to match what the
Class::Load code is trying to check.